requirements.txt
django
djangorestframework
mysqlclient==1.3.14
